Make logicalFromString() to use in xls and xlsx
